+++ app/grouping/create/components/inputSection/DateSection.server.tsx
@@ -4,7 +4,7 @@ import Calendar from '@/components/common/Calendar';
 import { BottomUpModal } from '@/components/common/Modal';
 import { DivisionSpacing } from '@/components/common/Spacing';
 import { TimeSwipePicker } from '@/components/common/SwipePicker';
-import useModalState from '@/store/useModals';
+import useModalState from '@/store/useModalStore';
 import { UseFormSetValue } from 'react-hook-form';
 
 import type { CreateMeetingRequestType, ModalNameType } from '../../type';
