+++ app/grouping/create/components/inputSection/NumberSection.server.tsx
@@ -2,7 +2,7 @@ import InputArea from './InputArea.server';
 import { BottomFixedButton } from '@/components/common/Button';
 import { BottomUpModal } from '@/components/common/Modal';
 import { NumberSwipePicker } from '@/components/common/SwipePicker';
-import useModalState from '@/store/useModals';
+import useModalState from '@/store/useModalStore';
 import { UseFormSetValue } from 'react-hook-form';
 
 import type { CreateMeetingRequestType } from '../../type';
